Trekker ut harFullførtFjerdetrinn til egen klasse for å kunne teste
